pyproject.toml
[project]
name = "image-describer"
version = "0.1.0"
description = "CLI tool to generate image descriptions with Ollama Vision"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
]

[project.scripts]
image-describer = "image_describer.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_describer"]
